+++ CHANGELOG.md
@@ -2,13 +2,14 @@
 
 ## v1.5.0
 - Externalize variable retrieval through callback to support more usage scenarios like external name normalization.
+- Rename `parseVariables` to `loadVariables` and add logs.
 
 ## v1.4.0
 - Add support for glob pattern in variable file paths ([#4](https://github.com/qetza/replacetokens/issues/4)).
 - Add support for YAML variable files ([#5](https://github.com/qetza/replacetokens/issues/5)).
 - Add support for JSON with comments in variables ([#6](https://github.com/qetza/replacetokens/issues/6)).
 - Add `parseVariables` function to parse string variables (load, flatten & merge).
-- Removed `flattenAndMerge` function (use `parseVariables` instead).
+- Remove `flattenAndMerge` function (use `parseVariables` instead).
 
 ## v1.3.0
 - Make variable names case-insensitive.